+++ src/Utils/Mappings/CommandNameToEntityTypeCode.php
@@ -62,7 +62,8 @@ class CommandNameToEntityTypeCode extends \ArrayObject
                 CommandNames::IMPORT_CREATE_OK_FILE            => EntityTypeCodes::NONE,
                 CommandNames::IMPORT_CREATE_CONFIGURATION_FILE => EntityTypeCodes::NONE,
                 CommandNames::IMPORT_EXECUTE                   => EntityTypeCodes::NONE,
-                CommandNames::IMPORT_CONVERT_VALUE             => EntityTypeCodes::NONE
+                CommandNames::IMPORT_CONVERT_VALUE             => EntityTypeCodes::NONE,
+                CommandNames::IMPORT_DEBUG                     => EntityTypeCodes::DEBUG
             ),
             $mappings
         );